Little refactoring of the project structure
