+++ frontend/src/app/layout.tsx
@@ -4,7 +4,7 @@ import './globals.css'
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap/dist/css/bootstrap.css';
 import React from 'react'
-import AppNavbar from './navbar';
+import AppNavbar from '../components/navbar';
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
