+++ frontend/src/app/signin/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 import React from "react";
-import signIn from "@/firebase/auth/signin";
+import signIn from "@/services/firebase/auth/signin";
 import { useRouter } from 'next/navigation'
 import { useAuthContext } from "@/context/authContext";
