+++ frontend/src/context/authContext.tsx
@@ -4,7 +4,7 @@ import {
   getAuth,
   User,
 } from 'firebase/auth';
-import firebase_app from '@/firebase/config';
+import firebase_app from '@/services/firebase/config';
 
 export const auth = getAuth(firebase_app);
