+++ frontend/src/services/firebase/auth/logout.ts
